Update i18n tooling scripts
